fix: 엔티티 수정
