+++ src/main/java/com/sprint/mission/discodeit/entity/BinaryContent.java
@@ -1,29 +1,41 @@
 package com.sprint.mission.discodeit.entity;
 
+import com.sprint.mission.discodeit.entity.base.BaseEntity;
+import jakarta.persistence.Column;
+import jakarta.persistence.Entity;
+import jakarta.persistence.Lob;
+import jakarta.persistence.Table;
 import lombok.Getter;
 
 import java.io.Serializable;
 import java.time.Instant;
 import java.util.UUID;
 
+@Entity
 @Getter
-public class BinaryContent implements Serializable {
-    private static final long serialVersionUID = 1L;
-    private UUID id;
-    private Instant createdAt;
-    //
-    private String fileName;
-    private Long size;
-    private String contentType;
-    private byte[] bytes;
+@Table(name = "binary_contents")
+public class BinaryContent extends BaseEntity {
 
-    public BinaryContent(String fileName, Long size, String contentType, byte[] bytes) {
-        this.id = UUID.randomUUID();
-        this.createdAt = Instant.now();
-        //
-        this.fileName = fileName;
-        this.size = size;
-        this.contentType = contentType;
-        this.bytes = bytes;
-    }
+  @Column(name = "file_name", nullable = false, length = 255)
+  private String fileName;
+
+  @Column(name = "size", nullable = false)
+  private Long size;
+
+  @Column(name = "content_type", nullable = false, length = 100)
+  private String contentType;
+
+  @Column(name = "bytes")
+  @Lob
+  private byte[] bytes;
+
+  protected BinaryContent() {
+  }
+
+  public BinaryContent(String fileName, Long size, String contentType, byte[] bytes) {
+    this.fileName = fileName;
+    this.size = size;
+    this.contentType = contentType;
+    this.bytes = bytes;
+  }
 }
