--- src/main/java/com/sprint/mission/discodeit/entity/base/BinaryContent.java
@@ -1,36 +1,0 @@
-package com.sprint.mission.discodeit.entity.base;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.Lob;
-import jakarta.persistence.Table;
-import lombok.Getter;
-
-@Entity
-@Getter
-@Table(name = "binary_contents")
-public class BinaryContent extends BaseEntity {
-
-  @Column(name = "file_name", nullable = false, length = 255)
-  private String fileName;
-
-  @Column(name = "size", nullable = false)
-  private Long size;
-
-  @Column(name = "content_type", nullable = false, length = 100)
-  private String contentType;
-
-  @Column(name = "bytes")
-  @Lob
-  private byte[] bytes;
-
-  protected BinaryContent() {
-  }
-
-  public BinaryContent(String fileName, Long size, String contentType, byte[] bytes) {
-    this.fileName = fileName;
-    this.size = size;
-    this.contentType = contentType;
-    this.bytes = bytes;
-  }
-}
