--- src/main/java/com/sprint/mission/discodeit/entity/base/Message.java
@@ -1,48 +1,0 @@
-package com.sprint.mission.discodeit.entity.base;
-
-import jakarta.persistence.Column;
-import jakarta.persistence.ConstraintMode;
-import jakarta.persistence.Entity;
-import jakarta.persistence.ForeignKey;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.JoinTable;
-import jakarta.persistence.ManyToMany;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.Table;
-import java.util.List;
-import lombok.Getter;
-
-@Entity
-@Getter
-@Table(name = "messages")
-public class Message extends BaseUpdatableEntity {
-
-  @Column(name = "content", columnDefinition = "TEXT")
-  private String content;
-
-  @ManyToOne(optional = false)
-  @JoinColumn(name = "channel_id", nullable = false, foreignKey = @ForeignKey(ConstraintMode.CONSTRAINT))
-  private Channel channel;
-
-  @ManyToOne
-  @JoinColumn(name = "author_id", foreignKey = @ForeignKey(ConstraintMode.CONSTRAINT))
-  private User author;
-
-  @ManyToMany
-  @JoinTable(
-      name = "message_attachments",
-      joinColumns = @JoinColumn(name = "message_id"),
-      inverseJoinColumns = @JoinColumn(name = "attachment_id")
-  )
-  private List<BinaryContent> attachments;
-
-  protected Message() {
-  }
-
-  public Message(String content, Channel channel, User author, List<BinaryContent> attachments) {
-    this.content = content;
-    this.channel = channel;
-    this.author = author;
-    this.attachments = attachments;
-  }
-}
