+++ src/main/java/com/sprint/mission/discodeit/repository/UserRepository.java
@@ -5,14 +5,13 @@ import com.sprint.mission.discodeit.entity.User;
 import java.util.List;
 import java.util.Optional;
 import java.util.UUID;
+import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository {
-    User save(User user);
-    Optional<User> findById(UUID id);
-    Optional<User> findByUsername(String username);
-    List<User> findAll();
-    boolean existsById(UUID id);
-    void deleteById(UUID id);
-    boolean existsByEmail(String email);
-    boolean existsByUsername(String username);
+public interface UserRepository extends JpaRepository<User, UUID> {
+
+  Optional<User> findByUsername(String username);
+
+  boolean existsByEmail(String email);
+
+  boolean existsByUsername(String username);
 }
